changed Graphics3DDraw#glTranslatef(float, float, int) to Graphics3DDraw#glTranslatef(float, float, float)

+++ org.eclipse.draw3d.graphics3d/src/java/org/eclipse/draw3d/graphics3d/Graphics3DDraw.java
@@ -410,7 +410,7 @@ public interface Graphics3DDraw {
 
 	public abstract void glLineStipple(int factor, short pattern);
 
-	public abstract void glTranslatef(float x, float y, int z);
+	public abstract void glTranslatef(float x, float y, float z);
 
 	public abstract String glGetString(int name);
